requirements.txt
fabric~=2.7.1
pyyaml~=6.0
keyring~=23.9.3
invoke~=1.7.3
netifaces~=0.11.0
paramiko~=2.12.0
mergedeep~=1.3.4
sentinel~=1.0.0
requests